refactor(state): keep lease repair helpers internal
